type=rule ruletype=any | introduce all filter/actions related to schedule for DoS/Pbf/Qos Rules
