+++ CHANGELOG.txt
@@ -3,6 +3,7 @@ CHANGELOG
 2.0.58
 UTIL:
 * type=rule - improve filters for CI/CD
+* type=rule ruletype=any | introduce all filter/actions related to schedule for DoS/Pbf/Qos Rules
 
 BUGFIX:
 * type=rule/schedule | bugfix for is.expired
