+++ lib/pan_php_framework.php
@@ -310,6 +310,7 @@ require_once $basedir . '/rule-classes/Rule.php';
 require_once $basedir . '/rule-classes/trait/NegatableRule.php';
 require_once $basedir . '/rule-classes/trait/RulewithLogging.php';
 require_once $basedir . '/rule-classes/RuleWithUserID.php';
+require_once $basedir . '/rule-classes/RuleWithSchedule.php';
 require_once $basedir . '/rule-classes/SecurityRule.php';
 require_once $basedir . '/rule-classes/NatRule.php';
 require_once $basedir . '/rule-classes/DecryptionRule.php';
